Code Optimizations
* Casting double to float is a free cost operation generating no memory traffic, it allow to avoid overflow;
* Cast size_t to int isn't a problem in the fact that all changements are less that INT_MAX (32.767(2^(15)-1);
* It may be more suitable to make thrown objets as const, the compiler will optimize it to be a direct reference;
* Dynamic arrays should be deleted with the `delete[]` operator, otherwise it could lead to memory overflow;
* All struct members must be initialized;
* Be sure to avoid arithmetics overflow in computing instructions;

+++ Illusion/IllusionEngine/include/ecs/Component.h
@@ -146,7 +146,7 @@ namespace illusion::ecs {
 		 * @return	le nombre d'éléments que compose le component
 		 */
 		inline i32 size() {
-			return ToEntity.size();
+			return static_cast<i32>(ToEntity.size());
 		}
 
 	protected:
