+++ Illusion/IllusionEngine/src/ecs/Component.cpp
@@ -53,7 +53,7 @@ namespace illusion::ecs {
 
 	void Component::OnEntityCreate(entity_id id) {
 		id::id_type index = id::Index(id);
-		if (index >= ToData.size()) ToData.resize(index + 1);
+		if (index >= ToData.size()) ToData.resize(static_cast<size_t>(index) + 1);
 		ToData[index] = component_id{ id::invalid_id };
 	}
